# api/api.yml
swagger: "2.0"
info:
  title: CtoFunds REST API
  description: |
    定义CtoFunds服务端的API，REST风格，包括请求响应URI、状态和报文格式
  version: "1.0"
  contact:
    name: 唐睿
    email: "tangrui@ctofunds.com"
host: www.ctofunds.com
schemes:
  - https
  - http
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  token:
    description: 通过登录/login得到的JWT token
    type: apiKey
    name: X-AUTH-TOKEN
    in: header
security:
  - token: []
paths:
...
  /startups/:
    get:
      summary: 查看创业公司列表
      description: |
        查看所有创业公司，支持分页和过滤器
      parameters:
        - name: page
          in: path
          description: 当前页码
          required: true
          type: integer
          minimum: 1
        - name: pageSize
          in: path
          description: 每页项数
          required: true
          type: integer
          minimum: 1
        - name: name
          in: path
          description: 创业公司名称
          required: false
          type: string
      tags:
        - startup
      responses:
        200:
          description: 创业公司列表
          schema:
            properties:
              total:
                description: 创业公司总数
                type: integer
                minimum: 0
              page:
                description: 当前第几页
                type: integer
                minimum: 1
              pageSize:
                description: 每页最大的创业公司数
                type: integer
                minimum: 1
              data:
                type: array
                items:
                  $ref: '#/definitions/Startup'
    post:
      summary: 创建创业公司
      description: |
        根据请求创建创业公司
      tags:
        - startup
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Startup'
      responses:
        201:
          description: 成功创建创业公司
          schema:
            $ref: '#/definitions/Startup'
        400: *400
        401: *401
        403: *403
        500: *500
  /startups/{id}/:
    get:
      summary: 查询创业公司
      description: |
        得到指定创业公司的具体信息
      tags:
        - startup
      parameters:
        - &StartupIdPara
          name: id
          in: path
          description: 创业公司ID
          required: true
          type: integer
          format: int64
      tags:
        - startup
      responses:
        200:
          description: 成功查询创业公司
          schema:
            $ref: '#/definitions/Startup'
        400: *400
        401: *401
        403: *403
        404: *404
        500: *500
    delete:
      summary: 删除创业公司
      description: |
        删除指定创业公司，后端采用懒删除，只是将状态设成已删除，数据将根据一定物理策略删除
      tags:
        - startup
      parameters:
        - *StartupIdPara
      responses:
        200:
          description: 成功删除创业公司
          schema:
            $ref: '#/definitions/Startup'
        400: *400
        401: *401
        403: *403
        404: *404
        500: *500
    put:
      summary: 更新创业公司
      description: |
        根据请求中指定要更新的属性对创业公司进行更新
      tags:
        - startup
      parameters:
        - *StartupIdPara
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateStartupRequest'
      responses:
        200:
          description: 成功更新创业公司
          schema:
            $ref: '#/definitions/Startup'
        400: *400
        401: *401
        403: *403
        404: *404
        500: *500
  /founders/:
    get:
      summary: 查看创始人列表
      description: |
        查看所有创始人，支持分页和过滤器
      parameters:
        - name: page
          in: path
          description: 当前页码
          required: true
          type: integer
          minimum: 1
        - name: pageSize
          in: path
          description: 每页项数
          required: true
          type: integer
          minimum: 1
        - name: name
          in: path
          description: 创始人名称
          required: false
          type: string
        - name: startupId
          in: path
          description: 创业公司ID
          required: false
          type: integer
          format: int64
      tags:
        - founder
      responses:
        200:
          description: 创始人列表
          schema:
            properties:
              total:
                description: 创始人总数
                type: integer
                minimum: 0
              page:
                description: 当前第几页
                type: integer
                minimum: 1
              pageSize:
                description: 每页最大的创始人数
                type: integer
                minimum: 1
              data:
                type: array
                items:
                  $ref: '#/definitions/Founder'
    post:
      summary: 创建创始人
      description: |
        根据请求创建创始人
      tags:
        - founder
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Founder'
      responses:
        201:
          description: 成功创建创始人
          schema:
            $ref: '#/definitions/Founder'
        400: *400
        401: *401
        403: *403
        500: *500
  /founders/{id}/:
    get:
      summary: 查询创始人
      description: |
        得到指定创始人的具体信息
      tags:
        - founder
      parameters:
        - &FounderIdPara
          name: id
          in: path
          description: 创始人ID
          required: true
          type: integer
          format: int64
      tags:
        - founder
      responses:
        200:
          description: 成功查询创始人
          schema:
            $ref: '#/definitions/Founder'
        400: *400
        401: *401
        403: *403
        404: *404
        500: *500
    delete:
      summary: 删除创始人
      description: |
        删除指定创始人，后端采用懒删除，只是将状态设成已删除，数据将根据一定物理策略删除
      tags:
        - founder
      parameters:
        - *FounderIdPara
      responses:
        200:
          description: 成功删除创始人
          schema:
            $ref: '#/definitions/Founder'
        400: *400
        401: *401
        403: *403
        404: *404
        500: *500
    put:
      summary: 更新创始人
      description: |
        根据请求中指定要更新的属性对创始人进行更新
      tags:
        - founder
      parameters:
        - *FounderIdPara
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateFounderRequest'
      responses:
        200:
          description: 成功更新创始人
          schema:
            $ref: '#/definitions/Founder'
        400: *400
        401: *401
        403: *403
        404: *404
        500: *500
...
  /comments/:
    get:
      summary: 查看评论列表
      description: |
        查看所有评论，支持分页和过滤器
      parameters:
        - name: page
          in: path
          description: 当前页码
          required: true
          type: integer
          minimum: 1
        - name: pageSize
          in: path
          description: 每页项数
          required: true
          type: integer
          minimum: 1
        - name: userId
          in: path
          description: 评论发布者ID
          required: false
          type: integer
          format: int64
        - name: topicId
          in: path
          description: 评论的主题ID
          required: false
          type: integer
          format: int64
        - name: topicId
          in: path
          description: 评论的回复ID
          required: false
          type: integer
          format: int64
      tags:
        - comment
      responses:
        200:
          description: 评论列表
          schema:
            properties:
              total:
                description: 评论总数
                type: integer
                minimum: 0
              page:
                description: 当前第几页
                type: integer
                minimum: 1
              pageSize:
                description: 每页最大的评论数
                type: integer
                minimum: 1
              data:
                type: array
                items:
                  $ref: '#/definitions/Comment'
    post:
      summary: 创建评论
      description: |
        根据请求创建评论
      tags:
        - comment
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Comment'
      responses:
        201:
          description: 成功创建评论
          schema:
            $ref: '#/definitions/Comment'
        400: *400
        401: *401
        403: *403
        500: *500
  /comments/{id}/:
    get:
      summary: 查询评论
      description: |
        得到指定评论的具体信息
      tags:
        - comment
      parameters:
        - &CommentIdPara
          name: id
          in: path
          description: 评论ID
          required: true
          type: integer
          format: int64
      tags:
        - comment
      responses:
        200:
          description: 成功查询评论
          schema:
            $ref: '#/definitions/Comment'
        404: *404
    delete:
      summary: 删除评论
      description: |
        删除指定评论，后端采用懒删除，只是将状态设成已删除，数据将根据一定物理策略删除
      tags:
        - comment
      parameters:
        - *CommentIdPara
      responses:
        200:
          description: 成功删除评论
          schema:
            $ref: '#/definitions/Comment'
        400: *400
        401: *401
        403: *403
        404: *404
        500: *500
    put:
      summary: 更新评论
      description: |
        根据请求中指定要更新的属性对评论进行更新
      tags:
        - comment
      parameters:
        - *CommentIdPara
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateCommentRequest'
      responses:
        200:
          description: 成功更新评论
          schema:
            $ref: '#/definitions/Comment'
        400: *400
        401: *401
        403: *403
        404: *404
        500: *500
definitions:
...
  Startup:
    description:
      表示创业公司资源
    required:
      - name
    properties:
      id:
        description: 创业公司ID，由后端自动生成，创建创业公司时不需要
        type: integer
        format: int64
      name:
        description: 创业公司名称
        type: string
      corpName:
        description: 创业公司注册名称
        type: string
      startYear:
        description: 创业公司创始年份
        type: integer
      startMonth:
        description: 创业公司创始月份
        type: integer
      productName:
        description: 创业公司产品名称
        type: string
      productDesc:
        description: 创业公司产品描述
        type: string
      location:
        description: 创业公司地址
        type: string
      scope:
        description: 创业公司经营范围
        type: string
      phase:
        description: 创业公司融资阶段
        type: string
      intro:
        description: 创业公司介绍
        type: string
      corpIdPhoto:
        description: 创业公司照片地址
        type: string
      logo:
        description: 创业公司Logo图片地址
        type: string
      adminId:
        description: 创业公司管理员ID
        type: integer
        format: int64
      createdAt:
        description: 创建时间
        type: string
        format: date-time
      modifiedAt:
        description: 最近一次修改时间
        type: string
        format: date-time
  Founder:
    description:
      表示创业人资源
    required:
      - name
    properties:
      id:
        description: 创业人ID，由后端自动生成，创建创业人时不需要
        type: integer
        format: int64
      name:
        description: 创始人姓名
        type: string
      position:
        description: 创始人职位
        type: string
      email:
        description: 创始人邮件地址
        type: string
      intro:
        description: 创始人介绍
        type: string
      startupId:
        description: 创业公司ID
        type: integer
        format: int64
      createdAt:
        description: 创建时间
        type: string
        format: date-time
      modifiedAt:
        description: 最近一次修改时间
        type: string
        format: date-time
...
  Comment:
    description:
      表示评论资源
    required:
      - userId
    properties:
      id:
        description: 评论ID，由后端自动生成，创建时不需要
        type: integer
        format: int64
      userId:
        description: 评论发布者
        type: integer
        format: int64
      content:
        description: 评论内容
        type: string
      topicId:
        description: 评论主题的ID
        type: integer
        format: int64
      replyId:
        description: 评论主题的ID
        type: integer
        format: int64
      createdAt:
        description: 创建时间
        type: string
        format: date-time
      modifiedAt:
        description: 最近一次修改时间
        type: string
        format: date-time
...